feat(transactions): filter transactions by id

+++ account_api/account_api/transactions/schemas.py
@@ -10,7 +10,7 @@ class OutMixIn(BaseModel):
 
 class TransactionIn(BaseModel):
     type: Annotated[str, Field(description="Tipo de transação", max_length=20, nullable=False)]
-    value: Annotated[int, Field(description="Valor da transação", gt=0, nullable=False)]
+    value: Annotated[float, Field(description="Valor da transação", gt=0, nullable=False)]
     client: Annotated[ClientIn, Field(description="Nome do cliente", nullable=False)]
 
 class TransactionOut(TransactionIn, OutMixIn):
